--- apps/web/modules/feature-opt-in/hooks/index.ts
@@ -1,10 +1,0 @@
-export type {
-  NormalizedFeature,
-  ToggleLabels,
-  UseFeatureOptInResult,
-} from "@calcom/features/feature-opt-in/types";
-export type { UseFeatureOptInBannerResult } from "./useFeatureOptInBanner";
-export { useFeatureOptInBanner } from "./useFeatureOptInBanner";
-export { useOrganizationFeatureOptIn } from "./useOrganizationFeatureOptIn";
-export { useTeamFeatureOptIn } from "./useTeamFeatureOptIn";
-export { useUserFeatureOptIn } from "./useUserFeatureOptIn";
