+++ apps/web/modules/settings/teams/[id]/features-view.tsx
@@ -1,12 +1,11 @@
 "use client";
 
-import type { ReactElement } from "react";
-
 import { FeaturesSettings } from "@calcom/features/feature-opt-in/components/FeaturesSettings";
 import SettingsHeader from "@calcom/features/settings/appDir/SettingsHeader";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
+import type { ReactElement } from "react";
 
-import { useTeamFeatureOptIn } from "~/feature-opt-in/hooks";
+import { useTeamFeatureOptIn } from "~/feature-opt-in/hooks/useTeamFeatureOptIn";
 
 interface TeamFeaturesViewProps {
   teamId: number;
